TINY-7795: circular dependency issue 2
