+++ modules/tinymce/src/core/main/ts/delete/CefDelete.ts
@@ -5,7 +5,7 @@ import Editor from '../api/Editor';
 import CaretPosition from '../caret/CaretPosition';
 import * as CefUtils from '../dom/CefUtils';
 import * as NodeType from '../dom/NodeType';
-import { isCefAtEdgeSelected } from '../keyboard/NavigationUtils';
+import { isCefAtEdgeSelected } from '../keyboard/CefUtils';
 import * as CefDeleteAction from './CefDeleteAction';
 import * as DeleteElement from './DeleteElement';
 import * as DeleteUtils from './DeleteUtils';
